requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
WTForms==3.1.1
Werkzeug==3.0.1
SQLAlchemy==2.0.23
email-validator==2.3.0

## Indirect dependencies that installed automatically
    blinker==1.9.0 
    click==8.3.1 
    colorama==0.4.6 
    greenlet==3.3.0 
    idna==3.11 
    itsdangerous==2.2.0 
    Jinja2==3.1.6 
    MarkupSafe==3.0.3 
    typing_extensions==4.15.0 
